fix(updates): finish startup after managed nodepack repairs
