+++ ARCHITECTURE_DEBT.toml
@@ -65,7 +65,7 @@ next_extraction = "Extract warmup and hydration task ownership from shell launch
 id = "SS-ARCH-006"
 owner = "startup managed-ready runtime"
 paths = ["substitute/app/bootstrap/startup_managed_ready_runtime.py"]
-fingerprint = "sha256:469638d0c92c953e1a9718d02147890785db14280788fc1de09558fd81e29e61"
+fingerprint = "sha256:d20ca2a30a738ec78583bb93b7b0a2f3976260defa4925cedcd41217af567a7d"
 issue = "chore:SS-ARCH-006"
 review_by = 2026-12-15
 responsibilities = [
