+++ substitute/app/bootstrap/startup_managed_ready_state.py
@@ -35,6 +35,8 @@ from substitute.app.bootstrap.startup_model_metadata import (
 )
 from substitute.app.bootstrap.startup_readiness_controller import (
     StartupReadinessControllerState,
+)
+from substitute.app.bootstrap.startup_readiness_resources import (
     StartupReadinessStarter,
 )
 from substitute.app.bootstrap.startup_warmup_controller import StartupWarmupState
