+++ tests/app/bootstrap/startup_managed_ready_runtime/test_runtime_lifecycle.py
@@ -67,6 +67,8 @@ from substitute.app.bootstrap.ready_shell_trace_fields import (
 from substitute.app.bootstrap.startup_readiness_controller import (
     StartupReadinessController,
     StartupReadinessControllerState,
+)
+from substitute.app.bootstrap.startup_readiness_resources import (
     StartupReadinessStarter,
 )
 from substitute.app.bootstrap.startup_probe_tasks import RuntimeCompatibilityProbeResult
